+++ src/ResultHandler/Implementations/Success/SuccessResult.cs
@@ -1,7 +1,5 @@
-﻿using System.Net;
-using ResultHandler.Core.Base;
+﻿using ResultHandler.Core.Base;
 using ResultHandler.Core.Enums;
-using ResultHandler.Mapping;
 
 namespace ResultHandler.Implementations.Success;
 
@@ -21,12 +19,6 @@ public class SuccessResult : OperationResult
 
     public SuccessResult(string title, ResultStatus status)
         : base(true, status, title)
-    {
-    }
-
-    [Obsolete("Use SuccessResult(string title, ResultStatus status) instead.")]
-    public SuccessResult(string statusMessage, HttpStatusCode statusCode)
-        : base(true, statusCode.ToResultStatus(), statusMessage)
     {
     }
 }